+++ packages/create-ima-app/template/.eslintrc.js
@@ -4,6 +4,8 @@ module.exports = {
     'eslint:recommended',
     'plugin:react/recommended',
     'plugin:react/jsx-runtime',
+    'plugin:jest/recommended',
+    'plugin:jest/style',
     'plugin:prettier/recommended',
   ],
   rules: {