+++ packages/dev-utils/src/compileErrorParser/parsers/swcLoaderErrorParser.ts
@@ -1,8 +1,8 @@
 import stripAnsi from 'strip-ansi';
 import { StatsError } from 'webpack';
 
-import { RE_VALID_FRAME_FIREFOX } from '../../helpers';
 import { RE_FILE_PATH_REGEX, CompileError } from './parserUtils';
+import { RE_VALID_FRAME_FIREFOX } from '../../helpers';
 
 const RE_SWC_LINE_NUMBER = /(\d+) │/;