+++ utils/tests/testRunner.sh
@@ -7,27 +7,44 @@ PARALLEL_TEST_CONNECTIONS=300
 SKELETON_URL="https://github.com/seznam/IMA.js-skeleton.git"
 NPM_LOCAL_REGISTRY_URL_NO_PROTOCOL="localhost:4873"
 NPM_LOCAL_REGISTRY_URL="http://${NPM_LOCAL_REGISTRY_URL_NO_PROTOCOL}/"
+ROOT_DIR_IMA=`pwd`
+
+cd ..
 
 ROOT_DIR=`pwd`
-CREATE_IMA_APP_DIR="$ROOT_DIR/packages/create-ima-app"
+ROOT_DIR_IMA_APP="$ROOT_DIR/ima-app"
+
+cd "$ROOT_DIR_IMA"
+
+CREATE_IMA_APP_DIR="$ROOT_DIR_IMA/packages/create-ima-app"
 PACKAGE_VERSION="0.0.0-next"
-PACKAGES="cli core dev-utils error-overlay helpers hmr-client server react-page-renderer"
+PACKAGES="cli core create-ima-app dev-utils error-overlay helpers hmr-client server react-page-renderer"
 
 # Setup local registry
-node_modules/.bin/verdaccio -l "$NPM_LOCAL_REGISTRY_URL_NO_PROTOCOL" -c utils/benchmark/verdaccio_config.yml >/dev/null &
+node_modules/.bin/verdaccio -l "$NPM_LOCAL_REGISTRY_URL_NO_PROTOCOL" -c utils/tests/verdaccio_config.yml >/dev/null &
 NPM_LOCAL_REGISTRY_PID=$!
 
 npm config set "//$NPM_LOCAL_REGISTRY_URL_NO_PROTOCOL/:_authToken" "0"
 
 # Release ima packages to local registry
 for PACKAGE in $PACKAGES ; do
-    cd "$ROOT_DIR/packages/$PACKAGE"
+    cd "$ROOT_DIR_IMA/packages/$PACKAGE"
     echo "Working on $PACKAGE@$PACKAGE_VERSION"
     sed -i "s#\"version\":\s\".*\"#\"version\": \"$PACKAGE_VERSION\"#" package.json
 
     for PACKAGE_UPDATE in $PACKAGES ; do
         sed -i "s#\"@ima/$PACKAGE_UPDATE\":\s\".*\"#\"@ima/$PACKAGE_UPDATE\": \"$PACKAGE_VERSION\"#" package.json
+
+        if [[ "$PACKAGE" == "create-ima-app" ]]
+        then
+            sed -i "s#\"@ima/$PACKAGE_UPDATE\":\s\".*\"#\"@ima/$PACKAGE_UPDATE\": \"$PACKAGE_VERSION\"#" template/package.json
+        fi
     done
+
+    if [[ "$PACKAGE" == "create-ima-app" ]]
+    then
+        npx json -I -f template/package.json -e "this.overrides={\"@ima/cli\":\"0.0.0-next\",\"@ima/core\":\"0.0.0-next\",\"@ima/helpers\":\"0.0.0-next\"}"
+    fi
 
     sed -i "s#https://registry.npmjs.org/#${NPM_LOCAL_REGISTRY_URL}#" package.json
     npm publish
@@ -37,7 +54,7 @@ done
 npm config set @ima:registry=$NPM_LOCAL_REGISTRY_URL
 
 # Update create-ima-app versions
-cd "$ROOT_DIR"
+cd "$ROOT_DIR_IMA"
 node utils/version/create-ima-app-versions.js
 # Link current create-ima-app version to global scope
 cd "$CREATE_IMA_APP_DIR"
@@ -47,22 +64,13 @@ npm link
 cd "$ROOT_DIR"
 npx create-ima-app ima-app
 
-cd ./ima-app
+cd "$ROOT_DIR_IMA_APP"
 
 npm run build
-# Add customized environment configuration
-mv server/config/environment.js server/config/environment.orig.js
-cp "$ROOT_DIR/utils/benchmark/app/environment.js" server/config/environment.js
-NODE_ENV=prod node server/server.js &
-IMA_SKELETON_SERVER_PID=$!
 
-sleep 7
-
-# Run test
-cd "$ROOT_DIR"
-node_modules/.bin/autocannon -c $PARALLEL_TEST_CONNECTIONS --no-progress "$TARGET_WEB_URL"
+# Run tests
+source "$ROOT_DIR_IMA/utils/tests/createImaAppTests.sh"
 
 # Cleanup
 npm config delete @ima:registry
 kill $NPM_LOCAL_REGISTRY_PID
-kill $IMA_SKELETON_SERVER_PID